[feature/sdk36] 학식 카드 및 화면 코드 정리
